+++ change_log.md
@@ -19,8 +19,9 @@
     * nerr -> netError
     * check udp connection alive after connect()
     * use ReadFromUDP as the uniform UDP read interface
-	* close net.UDPConn when connected failed
-	* close net.Conn when connected failed
+    * close net.UDPConn when connected failed
+    * close net.Conn when connected failed
+    * Session::EndPointType() ->  Session::EndPoint()
 
 - 2018/03/17
     > improvement